Add routing and period selection links
