+++ app/constants.ts
@@ -1,0 +1,22 @@
+export const periods = {
+  "30D": {
+    label: "30D",
+    aggregation: "Day",
+    default: false,
+  },
+  "14D": {
+    label: "14D",
+    aggregation: "Day",
+    default: false,
+  },
+  "7D": {
+    label: "7D",
+    aggregation: "Day",
+    default: false,
+  },
+  "24H": {
+    label: "24H",
+    aggregation: "Hour",
+    default: true,
+  },
+};
